move post_content transform into own class
