+++ command.php
@@ -13,6 +13,7 @@ require_once( HH_HUGO_COMMAND_DIR . '/Markdownify/src/ConverterExtra.php' );
 
 // Migration classes
 require_once( HH_HUGO_COMMAND_DIR . '/inc/write-file.php' );
+require_once( HH_HUGO_COMMAND_DIR . '/inc/process-wp-post-content.php' );
 require_once( HH_HUGO_COMMAND_DIR . '/inc/migrate-post.php' );
