+++ tests/test-migrate-post.php
@@ -79,58 +79,6 @@ class HH_Hugo_Test_Migrate_Post extends WP_UnitTestCase {
 		$this->assertFalse( isset( $front_matter['description'] ) );
 	}
 
-	public function test_convert_link_embeds() {
-
-		// test input with http and https protocols for good measure
-		$input = "hi\n\nhttp://twitter.com/HacksHackers/status/804429947406286848\nhttps://twitter.com/botic/status/806587782705664005";
-		$output = "hi\n\n{{< tweet 804429947406286848 >}}\n{{< tweet 806587782705664005 >}}";
-		$this->assertEquals( $output, $this->migrator->convert_link_embeds( $input ) );
-
-		$expected = '{{< youtube tV175PYFgJg >}}';
-		$input = 'http://www.youtube.com/watch?v=tV175PYFgJg';
-		$this->assertEquals( $expected, $this->migrator->convert_link_embeds( $input ) );
-
-		// these should not be converted
-		$input = 'The url is http://twitter.com/HacksHackers/status/804429947406286848';
-		$this->assertEquals( $input, $this->migrator->convert_link_embeds( $input ) );
-		$input = 'http://not-twitter.com/HacksHackers/status/804429947406286848';
-		$this->assertEquals( $input, $this->migrator->convert_link_embeds( $input ) );
-	}
-
-	public function test_convert_shortcodes() {
-		// Note the secquence of the final char in the ID so you can debug
-		$testers = array(
-			array( '{{< youtube H2Ncxw1xfc1 >}}', '[youtube http://www.youtube.com/watch#!v=H2Ncxw1xfc1]' ),
-			array( '{{< youtube H2Ncxw1xfc2 >}}', '[youtube http://www.youtube.com/watch?v=H2Ncxw1xfc2]' ),
-			array( '{{< youtube H2Ncxw1xfc3 >}}', '[youtube http://www.youtube.com/watch?v=H2Ncxw1xfc3&w=320&h=240&fmt=1&rel=0&showsearch=1&hd=0]' ),
-			array( '{{< youtube jF-kELmmvg4 >}}', '[youtube http://www.youtube.com/v/jF-kELmmvg4]' ),
-			array( '{{< youtube 9FhMMmqzbD5 >}}', '[youtube http://www.youtube.com/v/9FhMMmqzbD5?fs=1&hl=en_US]' ),
-			array( '{{< youtube Rrohlqeir56 >}}', '[youtube http://youtu.be/Rrohlqeir56]' ),
-			array( '{{< vimeo 141351 >}}', '[vimeo 141351]' ),
-			array( '{{< vimeo 141352 >}}', '[vimeo http://vimeo.com/141352]' ),
-			array( '{{< vimeo 141353 >}}', '[vimeo 141353 h=500&w=350]' ),
-			array( '{{< vimeo 141354 >}}', '[vimeo id=141354 width=350 height=500]' ),
-		);
-		foreach ( $testers as $test ) {
-			$this->assertEquals( $test[0], $this->migrator->convert_shortcodes( $test[1] ) );
-		}
-	}
-
-	public function test_quote_shortcode() {
-		update_post_meta( $this->posts[0], 'quote', 'Lorem ipsum' );
-		$expected = '<blockquote>Lorem ipsum</blockquote>';
-		$content = sprintf( '[quote id="%d"]', $this->posts[0] );
-		$actual = $this->migrator->convert_shortcodes( $content );
-		$markdown = $this->migrator->transform_post_content( $content );
-
-		$this->assertEquals( $expected, $actual );
-		$this->assertEquals( '> Lorem ipsum', $markdown );
-
-		$expected = '';
-		$actual = $this->migrator->convert_shortcodes( sprintf( '[quote id="%d"]', $this->posts[1] ) );
-		$this->assertEquals( $expected, $actual );
-	}
-
 	public function test_filter_markdown() {
 		// empty line, line break in linked image, local domain
 		$input = "foo\n  \nbar[\n![img...[http://hackshackers.alley.dev/]";
@@ -168,49 +116,5 @@ class HH_Hugo_Test_Migrate_Post extends WP_UnitTestCase {
 		$this->assertEquals( 1, $this->migrator->count_tags( '<img src="http://foo.bar/img.png" />' ) );
 		$this->assertEquals( 2, $this->migrator->count_tags( '<p><span>hi</span></p>' ) );
 		$this->assertEquals( 0, $this->migrator->count_tags( 'hello world' ) );
-	}
-
-	public function test_strip_no_attr_tags() {
-		$expect = 'content of div';
-		$actual = $this->migrator->strip_no_attr_tags( '<div>content of div</div>' );
-		$this->assertEquals( $expect, $actual );
-
-		$expect = '<img src="http://foo.bar" />';
-		$actual = $this->migrator->strip_no_attr_tags( '<div><img src="http://foo.bar" /></div>' );
-		$this->assertEquals( $expect, $actual );
-
-		$expect = "content\nwith newline";
-		$actual = $this->migrator->strip_no_attr_tags( "<div>content\nwith newline</div>" );
-		$this->assertEquals( $expect, $actual );
-
-		$expect = '<div id="yup">content of div</div>';
-		$actual = $this->migrator->strip_no_attr_tags( '<div id="yup">content of div</div>' );
-		$this->assertEquals( $expect, $actual );
-
-		$expect = 'content of div';
-		$actual = $this->migrator->strip_no_attr_tags( '<span>content of div</span>' );
-		$this->assertEquals( $expect, $actual );
-
-		$expect = "content\nwith newline\n\ntest still";
-		$actual = $this->migrator->strip_no_attr_tags( "<div>content\nwith newline</div><div>test still</div>" );
-		$this->assertEquals( $expect, $actual );
-	}
-
-	public function test_hugo_figure() {
-		$input = $this->migrator->parse_hugo_figure_args( '<a href="http://twitpic.com/dcx0xd"><img src="//twitpic.com/show/thumb/dcx0xd.jpg" alt="@nypl_stereo is in da house #hhnyc " /></a> Hacks/Hackers needed 3D glasses to get the full Stereogranimator effect.' );
-		$expected = "{{< figure\n  link=\"http://twitpic.com/dcx0xd\"\n  src=\"//twitpic.com/show/thumb/dcx0xd.jpg\"\n  alt=\"@nypl_stereo is in da house #hhnyc\"\n  caption=\"Hacks/Hackers needed 3D glasses to get the full Stereogranimator effect.\"\n>}}\n\n";
-		$this->assertEquals( $expected, $this->migrator->hugo_figure( $input ) );
-
-		$input = $this->migrator->parse_hugo_figure_args( '<img src="//twitpic.com/show/thumb/dcx0xd.jpg" alt="@nypl_stereo is in da house #hhnyc " />Hacks/Hackers needed 3D glasses to get the full Stereogranimator effect.' );
-		$expected = "{{< figure\n  src=\"//twitpic.com/show/thumb/dcx0xd.jpg\"\n  alt=\"@nypl_stereo is in da house #hhnyc\"\n  caption=\"Hacks/Hackers needed 3D glasses to get the full Stereogranimator effect.\"\n>}}\n\n";
-		$this->assertEquals( $expected, $this->migrator->hugo_figure( $input ) );
-
-		$input = $this->migrator->parse_hugo_figure_args( '<img src="//twitpic.com/show/thumb/dcx0xd.jpg" alt="@nypl_stereo is in da house #hhnyc " />' );
-		$expected = "{{< figure\n  src=\"//twitpic.com/show/thumb/dcx0xd.jpg\"\n  alt=\"@nypl_stereo is in da house #hhnyc\"\n>}}\n\n";
-		$this->assertEquals( $expected, $this->migrator->hugo_figure( $input ) );
-
-		$input = $this->migrator->parse_hugo_figure_args( '<img alt="@nypl_stereo is in da house #hhnyc " />Hacks/Hackers needed 3D glasses to get the full Stereogranimator effect.' );
-		$expected = '';
-		$this->assertEquals( $expected, $this->migrator->hugo_figure( $input ) );
 	}
 }
